+++ commands.py
@@ -1,6 +1,6 @@
 head = "Hi! I am Alien Mask (tis a pun on someone's name...don't ask...my creator is rather terrible at jokes) \n\n\nThis is all the shit I can do so far: \n\n"
-commands = ["/meme <category> - Meme. Current Categories:\nregular (default)\ndank\nrelatable","/joke - Lame ass jokes","/image <search> - Gives an image(sometimes it might take a little while to get the image. Be patient)","/youtube <search term> - Gives a YouTube video","/wiki <search> - Searches wikipedia for stuff","/suggestion <text>- Suggest new features","/about - My Personal Data (Gives you the Zuck feels)"]
-comingSoon = ["Link shortner maybe idk","Dynamic categories for memes","Whatever my psycho creator decides","Some tech shit maybe"]
+commands = ["/meme <category> - Meme. Current Categories:\nregular (default)\ndank\nrelatable","/joke - Lame ass jokes","/image <search> - Gives an image(sometimes it might take a little while to get the image. Be patient)","/youtube <search term> - Gives a YouTube video","/wiki <search> - Searches wikipedia for stuff","/shorten <url> - URL Shortener","/suggestion <text>- Suggest new features","/about - My Personal Data (Gives you the Zuck feels)"]
+comingSoon = ["Whatever my psycho creator decides","Some tech shit maybe"]
 
 
 helpMsg = head + '\n'.join(commands)+'\n\nComing Soon:\n\n'+'\n'.join(comingSoon)